+++ web-app/packages/lib/src/modules/project/store.ts
@@ -8,7 +8,6 @@ import omit from 'lodash/omit'
 import Vue from 'vue'
 import { Module } from 'vuex'
 
-import { CustomError } from '@/common/errors'
 import { waitCursor } from '@/common/html_utils'
 import { filesDiff } from '@/common/mergin_utils'
 import { isAtLeastProjectRole, ProjectRole } from '@/common/permission_utils'
@@ -28,6 +27,7 @@ import {
   ReloadProjectAccessRequestPayload
 } from '@/modules/project/types'
 import { RootState } from '@/modules/types'
+import { AxiosError, AxiosResponse } from 'axios'
 
 interface File {
   chunks: any
@@ -391,17 +391,6 @@ const ProjectStore: Module<ProjectState, RootState> = {
           true
         )
         await dispatch('reloadProjectAccessRequests', payload)
-      } catch (err) {
-        const msg = err.response
-          ? err.response.data?.detail
-          : 'Failed to accept access request'
-        await dispatch(
-          'notificationModule/error',
-          { text: msg },
-          {
-            root: true
-          }
-        )
       } finally {
         waitCursor(false)
       }
@@ -535,7 +524,10 @@ const ProjectStore: Module<ProjectState, RootState> = {
       }
     },
 
-    async saveProjectSettings({ commit, dispatch }, payload) {
+    async saveProjectSettings(
+      { commit },
+      payload
+    ): Promise<AxiosResponse<ProjectDetail>> {
       const { namespace, newSettings, projectName } = payload
 
       try {
@@ -550,16 +542,8 @@ const ProjectStore: Module<ProjectState, RootState> = {
         commit('setProject', { project: saveProjectSettingsResponse.data })
         waitCursor(false)
         return saveProjectSettingsResponse
-      } catch (err) {
-        await dispatch(
-          'notificationModule/error',
-          {
-            text: err.response.data?.detail || 'Failed to save project settings'
-          },
-          { root: true }
-        )
+      } finally {
         waitCursor(false)
-        return new CustomError('Failed to save project settings', err?.response)
       }
     },
